+++ src/organizing/namespace-manner/decorators/autobind.ts
@@ -1,0 +1,15 @@
+namespace App{
+    // Autobind Decorator
+
+    export function Autobind(_: any, _2: string, descriptor: PropertyDescriptor){
+        const originalMethod = descriptor.value;
+        const ajustedDescriptor: PropertyDescriptor = {
+            configurable: true,
+            get(){
+                const boundFn = originalMethod.bind(this);
+                return boundFn;
+            }
+        };
+        return ajustedDescriptor;
+    }
+}
